refactor ordebook structure
